102-fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int fib[50];
+	unsigned long fib[50];
 	int i;
 
 	fib[0] = 1;
@@ -22,7 +22,7 @@ int main(void)
 
 	for (i = 0; i < 50; i++)
 	{
-		printf("%d", fib[i]);
+		printf("%lu", fib[i]);
 
 		if (i != 49)
 		{
